+++ public/app/app.ts
@@ -152,120 +152,6 @@ module App {
     //declare var google;
 
     // Start the application
-    angular.module('csWebApp', [
-        'csComp',
-        'ngSanitize',
-        'ui.bootstrap',
-        'ui.select',
-        'schemaForm',
-        'LocalStorageModule',
-        'angularUtils.directives.dirPagination',
-        'pascalprecht.translate',
-        'ngCookies', 'angularSpectrumColorpicker',
-        'wiz.markdown', 'ngAnimate',
-        'simCitySimDirective'
-    ])
-        .config(localStorageServiceProvider => {
-            localStorageServiceProvider.prefix = 'csMap';
-        })
-        .config(TimelineServiceProvider => {
-            TimelineServiceProvider.setTimelineOptions({
-                'width': '100%',
-                'eventMargin': 0,
-                'eventMarginAxis': 0,
-                'editable': false,
-                'layout': 'box'
-            });
-        })
-        .config(($locationProvider) => {
-            $locationProvider.html5Mode({
-                enabled: true,
-                requireBase: false
-            });
-        })
-        .config(($logProvider) => {
-            $logProvider.debugEnabled(true);
-        })
-        .config($translateProvider => {
-            // TODO ADD YOUR LOCAL TRANSLATIONS HERE, OR ALTERNATIVELY, CHECK OUT
-            // http://angular-translate.github.io/docs/#/guide/12_asynchronous-loading
-            // Translations.English.locale['MAP_LABEL'] = 'MY AWESOME MAP';
-            $translateProvider.translations('en', Translations.English.locale);
-            $translateProvider.translations('nl', Translations.Dutch.locale);
-            $translateProvider.preferredLanguage('en');
-            // Enable escaping of HTML
-            $translateProvider.useSanitizeValueStrategy('escape');
-        })
-        .config($languagesProvider => {
-            // Defines the GUI languages that you wish to use in your project.
-            // They will be available through a popup menu.
-            var languages = [];
-            languages.push({
-                key: 'en',
-                name: 'English',
-                img: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAALCAIAAAD5gJpuAAAAGXRFWHRTb2Z0d2FyZQBBZG9iZ' +
-                     'SBJbWFnZVJlYWR5ccllPAAAAflJREFUeNpinDRzn5qN3uFDt16+YWBg+Pv339+KGN0rbVP+//2rW5tf0Hfy/2+mr99+yKpy' +
-                     'Ol3Ydt8njEWIn8f9zj639NC7j78eP//8739GVUUhNUNuhl8//ysKeZrJ/v7z10Zb2PTQTIY1XZO2Xmfad+f7XgkXxuUrVB6' +
-                     'cjPVXef78JyMjA8PFuwyX7gAZj97+T2e9o3d4BWNp84K1NzubTjAB3fH0+fv6N3qP/ir9bW6ozNQCijB8/8zw/TuQ7r4/nd' +
-                     'vN5mZgkpPXiis3Pv34+ZPh5t23//79Rwehof/9/NDEgMrOXHvJcrllgpoRN8PFOwy/fzP8+gUlgZI/f/5xcPj/69e/37//A' +
-                     'UX+/mXRkN555gsOG2xt/5hZQMwF4r9///75++f3nz8nr75gSms82jfvQnT6zqvXPjC8e/srJQHo9P9fvwNtAHmG4f8zZ6dD' +
-                     'c3bIyM2LTNlsbtfM9OPHH3FhtqUz3eXX9H+cOy9ZMB2o6t/Pn0DHMPz/b+2wXGTvPlPGFxdcD+mZyjP8+8MUE6sa7a/xo6P' +
-                     'ykn1s4zdzIZ6///8zMGpKM2pKAB0jqy4UE7/msKat6Jw5mafrsxNtWZ6/fjvNLW29qv25pQd///n+5+/fxDDVbcc//P/zx/' +
-                     '36m5Ub9zL8+7t66yEROcHK7q5bldMBAgwADcRBCuVLfoEAAAAASUVORK5CYII='
-            });
-            languages.push({
-                key: 'nl',
-                name: 'Nederlands',
-                img: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAALCAIAAAD5gJpuAAAABGdBTUEAAK/INwWK6QAAABl0R' +
-                     'Vh0U29mdHdhcmUAQWRvYmUgSW1hZ2VSZWFkeXHJZTwAAAFXSURBVHjaYvzPgAD/UNlYEUAAkuTgCAAIBgJggq5VoAs1qM0v' +
-                     'dzmMz362vezjokxPGimkEQ5WoAQEKuK71zwCCKyB4c//J8+BShn+/vv/+w/D399AEox+//8FJH/9/wUU+cUoKw20ASCAWBh' +
-                     'EDf/LyDOw84BU//kDtgGI/oARmAHRDJQSFwVqAAggxo8fP/Ly8oKc9P8/AxjiAoyMjA8ePAAIIJZ///5BVIM0MOBWDpRlZP' +
-                     'zz5w9AALH8gyvCbz7QBrCJAAHEyKDYX15r/+j1199//v35++/Xn7+///77DST/wMl/f4Dk378K4jx7O2cABBALw7NP77/+e' +
-                     'v3xB0gOpOHfr99AdX9/gTVASKCGP//+8XCyMjC8AwggFoZfIHWSwpwQk4CW/AYjsKlA8u+ff////v33998/YPgBnQQQQIzA' +
-                     'aGNg+AVGf5AYf5BE/oCjGEIyAQQYAGvKZ4C6+xXRAAAAAElFTkSuQmCC'
-            });
-            $languagesProvider.setLanguages(languages);
-        })
-        .constant('webserviceUrl')
-        .controller('appCtrl', AppCtrl)
-        .run((SchemaService, messageBusService) => {
-            SchemaService.addCustomTypeHandler('point2d', (formItem, _schemaItem, _form) => {
-                formItem.type = 'template';
-                formItem.template =
-                    '<div ng-if="item.id">{{item.id}}</div>' +
-                    '<div ng-if="!item.id">({{item.x}}, {{item.y}})</div>';
-            });
-            SchemaService.addCustomTypeHandler('layer', function (formItem, _schemaItem, _form) {
-                formItem.type = 'array';
-
-                var layerId = formItem.layer;
-                var featureId = formItem.featureId;
-                var key = formItem.key;
-
-                messageBusService.subscribe('feature', function (title, feature) {
-                    var supportedOps = ['dropped', 'onFeatureUpdated', 'onFeatureRemoved'];
-
-                    if (feature && layerId === feature.layerId && featureId === feature.properties.featureTypeId
-                        && supportedOps.indexOf(title) >= 0) {
-                        var f = {
-                            id: feature.properties.Name,
-                            x: feature.geometry.coordinates[0],
-                            y: feature.geometry.coordinates[1]
-                        };
-
-                        switch (title) {
-                            case 'dropped':
-                                SchemaService.modelAddValue(key, 'list', f);
-                                break;
-                            case 'onFeatureUpdated':
-                                SchemaService.modelUpdateValue(key, 'list', f);
-                                break;
-                            case 'onFeatureRemoved':
-                                SchemaService.modelDeleteValue(key, 'list', f);
-                                break;
-                        }
-                    }
-                });
-            });
-        });
+    angular.module('csWebApp')
+        .controller('appCtrl', AppCtrl);
 }
