+++ public/index.html
@@ -53,16 +53,18 @@
 
     <script src="bower_components/angularUtils-pagination/dirPagination.js"></script>
     <script src="bower_components/csweb/dist-bower/csComp.js"></script>
+    <script src="app/index.js"></script>
     <script src="app/app.js"></script>
+    <script src="app/services/schema.js"></script>
+    <script src="app/services/simulation.js"></script>
+    <script src="app/simform/simform.directive.js"></script>
+    <script src="app/simlist/simlist.directive.js"></script>
     <script src="bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
     <script src="bower_components/angular-bootstrap/ui-bootstrap-tpls.min.js"></script>
-    <script src="bower_components/sim-city-sim-directive/dist/scripts/app.js"></script>
     <script src="bower_components/csweb/dist-bower/csTemplates.js"></script>
 
     <!--<script src="bower_components/jquery/dist/jquery.min.js"></script>-->
     <!--<script src="bower_components/angular/angular.min.js"></script>-->
-    <script src="bower_components/angular-resource/angular-resource.js"></script>
-    <script src="bower_components/angular-toastr/dist/angular-toastr.tpls.js"></script>
     <script src="bower_components/tv4/tv4.js"></script>
     <script src="bower_components/angular-sanitize/angular-sanitize.min.js"></script>
     <script src="bower_components/objectpath/lib/ObjectPath.js"></script>
@@ -169,7 +171,11 @@
                             </svg>
                         </li>
 
-                        <li class="lp-tab" data-target="#simdirective" data-toggle="tab">
+                        <li class="lp-tab" data-target="#simform" data-toggle="tab">
+                            <i class="fa fa-fire" aria-hidden="true"></i>
+                        </li>
+
+                        <li class="lp-tab" data-target="#simlist" data-toggle="tab">
                             <i class="fa fa-fire" aria-hidden="true"></i>
                         </li>
                     </ul>
@@ -184,11 +190,11 @@
                         <style-list class="tab-pane" id="styles"></style-list>
                         <dashboard-selection class="tab-pane" id="dashboardselection"></dashboard-selection>
                         <project-settings class="tab-pane" id="settings"></project-settings>
-                        <div style="overflow-y: auto; overflow-x: hidden; padding:5px; margin-top: -40px; margin-left:10px;" class="tab-pane ng-isolate-scope scrolling leftpanel-scroll" id="simdirective">
-                            <sim-form schemaurl="http://localhost/explore/simulate/matsim/latest"></sim-form>
+                        <div style="overflow-y: auto; overflow-x: hidden; padding:5px; margin-top: -40px; margin-left:10px;" class="tab-pane ng-isolate-scope scrolling leftpanel-scroll" id="simform">
+                            <sim-form sim-webservice-url="/explore"></sim-form>
                         </div>
-                        <div style="overflow-y: auto; overflow-x: hidden; padding:5px; margin-top: -40px; margin-left:10px;" class="tab-pane ng-isolate-scope scrolling leftpanel-scroll" id="simdirective">
-                            <sim-list schemaurl="http://localhost/explore/simulate/matsim/latest"></sim-list>
+                        <div style="overflow-y: auto; overflow-x: hidden; padding:5px; margin-top: -40px; margin-left:10px;" class="tab-pane ng-isolate-scope scrolling leftpanel-scroll" id="simlist">
+                            <sim-list sim-webservice-url="/explore" sim-name="matsim" sim-version="latest"></sim-list>
                         </div>
                     </div>
                 </div>
